third_party/blink/web_tests/http/tests/local/formdata/send-form-data-with-string-containing-null-expected.txt

CONSOLE WARNING: Synchronous XMLHttpRequest on the main thread is deprecated because of its detrimental effects to the end user's experience. For more help, check https://xhr.spec.whatwg.org/.
Test that we can send textarea data containing null characters.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS the string containing a null byte is echoed correctly.
PASS successfullyParsed is true

TEST COMPLETE
